feat: disable cookie/localstorage access under restricted sandboxes
When dash is embedded into an iframe with a sandbox attribute that only has allow-scripts, cookie/localstorage access is disabled and dash-core-components fails to load. As such, we need to restrict our cookie/localstorage usage by disabling functionality.

This patch removes the disabled functionality in a graceful manner, allowing dash-core-components to load in very restricted iframes.